refactor(processor): split report.go into focused files
Extract statistics types and generation into report_statistics.go
and markdown rendering into report_markdown.go. Pure file-split
with no logic changes — reduces report.go from 820 to 256 lines.

Closes #320
