+++ internal/processor/report.go
@@ -3,16 +3,11 @@ package processor
 import (
 	"encoding/json"
 	"fmt"
-	"maps"
-	"slices"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/EvilBit-Labs/opnDossier/internal/analysis"
 	common "github.com/EvilBit-Labs/opnDossier/pkg/model"
-	"github.com/nao1215/markdown"
 	"gopkg.in/yaml.v3"
 )
 
@@ -58,90 +53,6 @@ type ConfigInfo struct {
 	Theme string `json:"theme,omitempty"`
 	// DeviceType identifies the platform
 	DeviceType common.DeviceType `json:"deviceType,omitempty"`
-}
-
-// Statistics contains various statistics about the configuration.
-type Statistics struct {
-	// Interface statistics
-	TotalInterfaces  int                   `json:"totalInterfaces"`
-	InterfacesByType map[string]int        `json:"interfacesByType"`
-	InterfaceDetails []InterfaceStatistics `json:"interfaceDetails"`
-
-	// Firewall and NAT statistics
-	TotalFirewallRules int            `json:"totalFirewallRules"`
-	RulesByInterface   map[string]int `json:"rulesByInterface"`
-	RulesByType        map[string]int `json:"rulesByType"`
-	NATEntries         int            `json:"natEntries"`
-	NATMode            string         `json:"natMode"`
-
-	// Gateway statistics
-	TotalGateways      int `json:"totalGateways"`
-	TotalGatewayGroups int `json:"totalGatewayGroups"`
-
-	// DHCP statistics
-	DHCPScopes       int                   `json:"dhcpScopes"`
-	DHCPScopeDetails []DHCPScopeStatistics `json:"dhcpScopeDetails"`
-
-	// User and group statistics
-	TotalUsers    int            `json:"totalUsers"`
-	UsersByScope  map[string]int `json:"usersByScope"`
-	TotalGroups   int            `json:"totalGroups"`
-	GroupsByScope map[string]int `json:"groupsByScope"`
-
-	// Service statistics
-	EnabledServices []string            `json:"enabledServices"`
-	TotalServices   int                 `json:"totalServices"`
-	ServiceDetails  []ServiceStatistics `json:"serviceDetails"`
-
-	// System configuration statistics
-	SysctlSettings       int      `json:"sysctlSettings"`
-	LoadBalancerMonitors int      `json:"loadBalancerMonitors"`
-	SecurityFeatures     []string `json:"securityFeatures"`
-
-	// IDS/IPS statistics
-	IDSEnabled             bool     `json:"idsEnabled"`
-	IDSMode                string   `json:"idsMode"`
-	IDSMonitoredInterfaces []string `json:"idsMonitoredInterfaces,omitempty"`
-	IDSDetectionProfile    string   `json:"idsDetectionProfile,omitempty"`
-	IDSLoggingEnabled      bool     `json:"idsLoggingEnabled"`
-
-	// Summary counts for quick reference
-	Summary StatisticsSummary `json:"summary"`
-}
-
-// InterfaceStatistics contains detailed statistics for a single interface.
-type InterfaceStatistics struct {
-	Name        string `json:"name"`
-	Type        string `json:"type"`
-	Enabled     bool   `json:"enabled"`
-	HasIPv4     bool   `json:"hasIpv4"`
-	HasIPv6     bool   `json:"hasIpv6"`
-	HasDHCP     bool   `json:"hasDhcp"`
-	BlockPriv   bool   `json:"blockPriv"`
-	BlockBogons bool   `json:"blockBogons"`
-}
-
-// DHCPScopeStatistics contains statistics for DHCP scopes.
-type DHCPScopeStatistics struct {
-	Interface string `json:"interface"`
-	Enabled   bool   `json:"enabled"`
-	From      string `json:"from"`
-	To        string `json:"to"`
-}
-
-// ServiceStatistics contains statistics for individual services.
-type ServiceStatistics struct {
-	Name    string            `json:"name"`
-	Enabled bool              `json:"enabled"`
-	Details map[string]string `json:"details,omitempty"`
-}
-
-// StatisticsSummary provides high-level summary statistics.
-type StatisticsSummary struct {
-	TotalConfigItems    int  `json:"totalConfigItems"`
-	SecurityScore       int  `json:"securityScore"`
-	ConfigComplexity    int  `json:"configComplexity"`
-	HasSecurityFeatures bool `json:"hasSecurityFeatures"`
 }
 
 // Findings contains analysis findings categorized by severity and type.
@@ -339,482 +250,4 @@ func (r *Report) redactedCopyUnsafe() *Report {
 	}
 
 	return cp
-}
-
-// ToMarkdown returns the report formatted as Markdown using the markdown library.
-func (r *Report) ToMarkdown() string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	var buf strings.Builder
-	md := markdown.NewMarkdown(&buf)
-
-	r.addHeader(md)
-	r.addConfigInfo(md)
-
-	if r.Statistics != nil {
-		r.addStatistics(md)
-	}
-
-	r.addFindingsUnsafe(md)
-
-	if err := md.Build(); err != nil {
-		return "# OPNsense Configuration Analysis Report\n\nError generating report.\n"
-	}
-
-	return buf.String()
-}
-
-// Helper functions for Markdown generation
-
-// addHeader writes the top-level report title and generation timestamp to the markdown document.
-func (r *Report) addHeader(md *markdown.Markdown) {
-	md.H1("OPNsense Configuration Analysis Report").
-		PlainTextf("Generated: %s", r.GeneratedAt.Format(time.RFC3339)).
-		LF()
-}
-
-// addConfigInfo writes the configuration metadata section (hostname, domain, version, theme) to the markdown document.
-func (r *Report) addConfigInfo(md *markdown.Markdown) {
-	md.H2("Configuration Information")
-	configItems := []string{
-		fmt.Sprintf("%s: %s", markdown.Bold("Hostname"), r.ConfigInfo.Hostname),
-		fmt.Sprintf("%s: %s", markdown.Bold("Domain"), r.ConfigInfo.Domain),
-	}
-	if r.ConfigInfo.Version != "" {
-		configItems = append(configItems, fmt.Sprintf("%s: %s", markdown.Bold("Version"), r.ConfigInfo.Version))
-	}
-	if r.ConfigInfo.Theme != "" {
-		configItems = append(configItems, fmt.Sprintf("%s: %s", markdown.Bold("Theme"), r.ConfigInfo.Theme))
-	}
-	md.BulletList(configItems...)
-	md.LF()
-}
-
-// addStatistics writes the configuration statistics overview and summary sections to the markdown document.
-func (r *Report) addStatistics(md *markdown.Markdown) {
-	overviewItems := []string{
-		fmt.Sprintf("%s: %d", markdown.Bold("Total Interfaces"), r.Statistics.TotalInterfaces),
-		fmt.Sprintf("%s: %d", markdown.Bold("Firewall Rules"), r.Statistics.TotalFirewallRules),
-		fmt.Sprintf("%s: %d", markdown.Bold("NAT Entries"), r.Statistics.NATEntries),
-		fmt.Sprintf("%s: %d", markdown.Bold("DHCP Scopes"), r.Statistics.DHCPScopes),
-		fmt.Sprintf("%s: %d", markdown.Bold("Users"), r.Statistics.TotalUsers),
-		fmt.Sprintf("%s: %d", markdown.Bold("Groups"), r.Statistics.TotalGroups),
-		fmt.Sprintf("%s: %d", markdown.Bold("Services"), r.Statistics.TotalServices),
-		fmt.Sprintf("%s: %d", markdown.Bold("Sysctl Settings"), r.Statistics.SysctlSettings),
-	}
-
-	summaryItems := []string{
-		fmt.Sprintf("%s: %d", markdown.Bold("Total Configuration Items"), r.Statistics.Summary.TotalConfigItems),
-		fmt.Sprintf("%s: %d/100", markdown.Bold("Security Score"), r.Statistics.Summary.SecurityScore),
-		fmt.Sprintf("%s: %d/100", markdown.Bold("Configuration Complexity"), r.Statistics.Summary.ConfigComplexity),
-		fmt.Sprintf("%s: %t", markdown.Bold("Has Security Features"), r.Statistics.Summary.HasSecurityFeatures),
-	}
-
-	md.
-		H2("Configuration Statistics").
-		H3("Overview").
-		BulletList(overviewItems...).
-		LF().
-		H3("Summary Metrics").
-		BulletList(summaryItems...).
-		LF()
-
-	addInterfaceDetails(md, "Interface Details", r.Statistics.InterfaceDetails)
-	addStatisticsList(md, "Firewall Rules by Interface", r.Statistics.RulesByInterface, " rules")
-	addStatisticsList(md, "Firewall Rules by Type", r.Statistics.RulesByType, " rules")
-	addDHCPScopeDetails(md, "DHCP Scope Details", r.Statistics.DHCPScopeDetails)
-	addStatisticsList(md, "Users by Scope", r.Statistics.UsersByScope, " users")
-	addStatisticsList(md, "Groups by Scope", r.Statistics.GroupsByScope, " groups")
-
-	if len(r.Statistics.EnabledServices) > 0 {
-		md.H3("Enabled Services").
-			BulletList(r.Statistics.EnabledServices...).
-			LF()
-	}
-
-	if len(r.Statistics.ServiceDetails) > 0 {
-		md.H3("Service Details")
-		for _, service := range r.Statistics.ServiceDetails {
-			serviceItems := []string{
-				fmt.Sprintf("%s: %t", markdown.Bold("Enabled"), service.Enabled),
-			}
-			// Sort detail keys for deterministic output
-			detailKeys := slices.Sorted(maps.Keys(service.Details))
-			for _, k := range detailKeys {
-				serviceItems = append(serviceItems, fmt.Sprintf("%s: %s", markdown.Bold(k), service.Details[k]))
-			}
-
-			md.
-				H4(service.Name).
-				BulletList(serviceItems...).
-				LF()
-		}
-	}
-
-	if len(r.Statistics.SecurityFeatures) > 0 {
-		md.
-			H3("Security Features").
-			BulletList(r.Statistics.SecurityFeatures...).
-			LF()
-	}
-
-	if r.Statistics.NATMode != "" {
-		natItems := []string{
-			fmt.Sprintf("%s: %s", markdown.Bold("NAT Mode"), r.Statistics.NATMode),
-			fmt.Sprintf("%s: %d", markdown.Bold("NAT Entries"), r.Statistics.NATEntries),
-		}
-		md.
-			H3("NAT Configuration").
-			BulletList(natItems...).
-			LF()
-	}
-
-	if r.Statistics.LoadBalancerMonitors > 0 {
-		lbItems := []string{
-			fmt.Sprintf("%s: %d", markdown.Bold("Monitors"), r.Statistics.LoadBalancerMonitors),
-		}
-		md.
-			H3("Load Balancer").
-			BulletList(lbItems...).
-			LF()
-	}
-
-	if r.Statistics.IDSEnabled {
-		idsItems := []string{
-			markdown.Bold("Status") + ": Enabled",
-			fmt.Sprintf("%s: %s", markdown.Bold("Mode"), r.Statistics.IDSMode),
-		}
-		if len(r.Statistics.IDSMonitoredInterfaces) > 0 {
-			idsItems = append(idsItems, fmt.Sprintf("%s: %s",
-				markdown.Bold("Monitored Interfaces"),
-				strings.Join(r.Statistics.IDSMonitoredInterfaces, ", ")))
-		}
-		if r.Statistics.IDSDetectionProfile != "" {
-			idsItems = append(idsItems, fmt.Sprintf("%s: %s",
-				markdown.Bold("Detection Profile"), r.Statistics.IDSDetectionProfile))
-		}
-		idsItems = append(idsItems, fmt.Sprintf("%s: %t",
-			markdown.Bold("Logging Enabled"), r.Statistics.IDSLoggingEnabled))
-
-		md.
-			H3("IDS/IPS Configuration").
-			BulletList(idsItems...).
-			LF()
-	}
-}
-
-// addFindingsUnsafe adds findings to markdown. Caller must hold mu.
-func (r *Report) addFindingsUnsafe(md *markdown.Markdown) {
-	md.H2("Analysis Findings")
-	if r.totalFindingsUnsafe() == 0 {
-		md.
-			PlainText("No findings to report.").
-			LF()
-		return
-	}
-
-	md.
-		PlainText(fmt.Sprintf("Total findings: %d", r.totalFindingsUnsafe())).
-		LF()
-
-	r.addFindingsSection(md, "Critical", r.Findings.Critical)
-	r.addFindingsSection(md, "High", r.Findings.High)
-	r.addFindingsSection(md, "Medium", r.Findings.Medium)
-	r.addFindingsSection(md, "Low", r.Findings.Low)
-	r.addFindingsSection(md, "Informational", r.Findings.Info)
-}
-
-// addStatisticsList renders a sorted bullet list of named integer statistics under the given title,
-// appending suffix to each value.
-func addStatisticsList(md *markdown.Markdown, title string, stats map[string]int, suffix string) {
-	if len(stats) == 0 {
-		return
-	}
-	// Sort keys for deterministic output
-	keys := slices.Sorted(maps.Keys(stats))
-	items := make([]string, 0, len(keys))
-	for _, k := range keys {
-		items = append(items, fmt.Sprintf("%s: %d%s", markdown.Bold(k), stats[k], suffix))
-	}
-	md.
-		H3(title).
-		BulletList(items...).
-		LF()
-}
-
-// addInterfaceDetails renders a table of interface statistics (type, enabled state, addressing, blocking) under the given title.
-func addInterfaceDetails(md *markdown.Markdown, title string, details []InterfaceStatistics) {
-	if len(details) == 0 {
-		return
-	}
-	table := markdown.TableSet{
-		Header: []string{"Interface", "Type", "Enabled", "IPv4", "IPv6", "DHCP", "Block Private", "Block Bogons"},
-		Rows:   [][]string{},
-	}
-	for _, detail := range details {
-		table.Rows = append(table.Rows, []string{
-			detail.Name,
-			detail.Type,
-			strconv.FormatBool(detail.Enabled),
-			strconv.FormatBool(detail.HasIPv4),
-			strconv.FormatBool(detail.HasIPv6),
-			strconv.FormatBool(detail.HasDHCP),
-			strconv.FormatBool(detail.BlockPriv),
-			strconv.FormatBool(detail.BlockBogons),
-		})
-	}
-
-	md.
-		H3(title).
-		Table(table).
-		LF()
-}
-
-// addDHCPScopeDetails renders a table of DHCP scope statistics (interface, enabled state, address range) under the given title.
-func addDHCPScopeDetails(md *markdown.Markdown, title string, details []DHCPScopeStatistics) {
-	if len(details) == 0 {
-		return
-	}
-	table := markdown.TableSet{
-		Header: []string{"Interface", "Enabled", "Range Start", "Range End"},
-		Rows:   [][]string{},
-	}
-	for _, detail := range details {
-		table.Rows = append(table.Rows, []string{
-			detail.Interface,
-			strconv.FormatBool(detail.Enabled),
-			detail.From,
-			detail.To,
-		})
-	}
-	md.
-		H3(title).
-		Table(table).
-		LF()
-}
-
-// Summary returns a brief summary of the report.
-func (r *Report) Summary() string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	var buf strings.Builder
-	md := markdown.NewMarkdown(&buf)
-
-	// Title with hostname
-	hostname := r.ConfigInfo.Hostname
-	if r.ConfigInfo.Domain != "" {
-		hostname += "." + r.ConfigInfo.Domain
-	}
-	md.
-		PlainText("OPNsense Configuration Report for " + hostname).
-		LF()
-
-	// Configuration statistics
-	if r.Statistics != nil {
-		md.
-			PlainTextf("Configuration contains %d interfaces, %d firewall rules, %d users, and %d groups.",
-				r.Statistics.TotalInterfaces,
-				r.Statistics.TotalFirewallRules,
-				r.Statistics.TotalUsers,
-				r.Statistics.TotalGroups).
-			LF()
-	}
-
-	// Findings summary
-	totalFindings := r.totalFindingsUnsafe()
-	if totalFindings == 0 {
-		md.PlainText("No issues found in the configuration.")
-	} else {
-		parts := []string{}
-		if len(r.Findings.Critical) > 0 {
-			parts = append(parts, fmt.Sprintf("%d critical", len(r.Findings.Critical)))
-		}
-		if len(r.Findings.High) > 0 {
-			parts = append(parts, fmt.Sprintf("%d high", len(r.Findings.High)))
-		}
-		if len(r.Findings.Medium) > 0 {
-			parts = append(parts, fmt.Sprintf("%d medium", len(r.Findings.Medium)))
-		}
-		if len(r.Findings.Low) > 0 {
-			parts = append(parts, fmt.Sprintf("%d low", len(r.Findings.Low)))
-		}
-		if len(r.Findings.Info) > 0 {
-			parts = append(parts, fmt.Sprintf("%d info", len(r.Findings.Info)))
-		}
-
-		md.PlainTextf("Analysis found %d findings: %s.", totalFindings, strings.Join(parts, ", "))
-	}
-
-	if err := md.Build(); err != nil {
-		return fmt.Sprintf("Error generating summary: %v", err)
-	}
-
-	return buf.String()
-}
-
-// addFindingsSection adds a findings section using the markdown library.
-func (r *Report) addFindingsSection(md *markdown.Markdown, title string, findings []Finding) {
-	if len(findings) == 0 {
-		return
-	}
-
-	md.H3(fmt.Sprintf("%s (%d)", title, len(findings)))
-
-	for i, finding := range findings {
-		md.H4(fmt.Sprintf("%d. %s", i+1, finding.Title))
-
-		findingItems := []string{
-			fmt.Sprintf("%s: %s", markdown.Bold("Type"), finding.Type),
-		}
-
-		if finding.Component != "" {
-			findingItems = append(findingItems, fmt.Sprintf("%s: %s", markdown.Bold("Component"), finding.Component))
-		}
-
-		findingItems = append(findingItems, fmt.Sprintf("%s: %s", markdown.Bold("Description"), finding.Description))
-
-		if finding.Recommendation != "" {
-			findingItems = append(
-				findingItems,
-				fmt.Sprintf("%s: %s", markdown.Bold("Recommendation"), finding.Recommendation),
-			)
-		}
-
-		if finding.Reference != "" {
-			findingItems = append(findingItems, fmt.Sprintf("%s: %s", markdown.Bold("Reference"), finding.Reference))
-		}
-
-		md.
-			BulletList(findingItems...).
-			HorizontalRule().
-			LF()
-	}
-}
-
-// redactServiceDetails replaces sensitive values in the processor's ServiceDetails
-// with the redaction marker. The shared analysis.ComputeStatistics intentionally
-// returns raw values so callers can decide on redaction policy; the processor
-// always redacts before rendering or serializing.
-func redactServiceDetails(stats *Statistics) {
-	if stats == nil {
-		return
-	}
-
-	for i := range stats.ServiceDetails {
-		if stats.ServiceDetails[i].Name == analysis.ServiceNameSNMP && stats.ServiceDetails[i].Details != nil {
-			if _, ok := stats.ServiceDetails[i].Details["community"]; ok {
-				// Deep-copy the map to avoid mutating shared state.
-				stats.ServiceDetails[i].Details = maps.Clone(stats.ServiceDetails[i].Details)
-				stats.ServiceDetails[i].Details["community"] = redactedValue
-			}
-		}
-	}
-}
-
-// generateStatistics analyzes a device configuration and returns aggregated statistics.
-// It delegates to analysis.ComputeStatistics for shared logic, translates the result
-// into the processor's Statistics type, and adds IDS-specific fields.
-func generateStatistics(cfg *common.CommonDevice) *Statistics {
-	commonStats := analysis.ComputeStatistics(cfg)
-	stats := translateCommonStats(commonStats)
-
-	// IDS/IPS statistics — processor-specific fields not in common.Statistics
-	ids := cfg.IDS
-	if ids != nil && ids.Enabled {
-		stats.IDSEnabled = true
-		stats.IDSMonitoredInterfaces = ids.Interfaces
-		stats.IDSDetectionProfile = ids.Detect.Profile
-		stats.IDSLoggingEnabled = ids.SyslogEnabled || ids.SyslogEveEnabled
-
-		if ids.IPSMode {
-			stats.IDSMode = "IPS (Prevention)"
-		} else {
-			stats.IDSMode = "IDS (Detection Only)"
-		}
-	}
-
-	// Redact sensitive service details before the report can be rendered or serialized.
-	redactServiceDetails(stats)
-
-	return stats
-}
-
-// translateCommonStats converts a common.Statistics into a processor.Statistics
-// by copying all matching fields. The processor's Statistics type mirrors the
-// common type but adds IDS-specific fields.
-func translateCommonStats(cs *common.Statistics) *Statistics {
-	stats := &Statistics{
-		TotalInterfaces:  cs.TotalInterfaces,
-		InterfacesByType: cs.InterfacesByType,
-		RulesByInterface: cs.RulesByInterface,
-		RulesByType:      cs.RulesByType,
-
-		TotalFirewallRules: cs.TotalFirewallRules,
-		NATEntries:         cs.NATEntries,
-		NATMode:            cs.NATMode,
-
-		TotalGateways:      cs.TotalGateways,
-		TotalGatewayGroups: cs.TotalGatewayGroups,
-
-		DHCPScopes: cs.DHCPScopes,
-
-		TotalUsers:    cs.TotalUsers,
-		UsersByScope:  cs.UsersByScope,
-		TotalGroups:   cs.TotalGroups,
-		GroupsByScope: cs.GroupsByScope,
-
-		EnabledServices: cs.EnabledServices,
-		TotalServices:   cs.TotalServices,
-
-		SysctlSettings:       cs.SysctlSettings,
-		LoadBalancerMonitors: cs.LoadBalancerMonitors,
-		SecurityFeatures:     cs.SecurityFeatures,
-
-		Summary: StatisticsSummary{
-			TotalConfigItems:    cs.Summary.TotalConfigItems,
-			SecurityScore:       cs.Summary.SecurityScore,
-			ConfigComplexity:    cs.Summary.ConfigComplexity,
-			HasSecurityFeatures: cs.Summary.HasSecurityFeatures,
-		},
-	}
-
-	// Translate InterfaceDetails
-	stats.InterfaceDetails = make([]InterfaceStatistics, len(cs.InterfaceDetails))
-	for i, d := range cs.InterfaceDetails {
-		stats.InterfaceDetails[i] = InterfaceStatistics{
-			Name:        d.Name,
-			Type:        d.Type,
-			Enabled:     d.Enabled,
-			HasIPv4:     d.HasIPv4,
-			HasIPv6:     d.HasIPv6,
-			HasDHCP:     d.HasDHCP,
-			BlockPriv:   d.BlockPriv,
-			BlockBogons: d.BlockBogons,
-		}
-	}
-
-	// Translate DHCPScopeDetails
-	stats.DHCPScopeDetails = make([]DHCPScopeStatistics, len(cs.DHCPScopeDetails))
-	for i, d := range cs.DHCPScopeDetails {
-		stats.DHCPScopeDetails[i] = DHCPScopeStatistics{
-			Interface: d.Interface,
-			Enabled:   d.Enabled,
-			From:      d.From,
-			To:        d.To,
-		}
-	}
-
-	// Translate ServiceDetails
-	stats.ServiceDetails = make([]ServiceStatistics, len(cs.ServiceDetails))
-	for i, d := range cs.ServiceDetails {
-		stats.ServiceDetails[i] = ServiceStatistics{
-			Name:    d.Name,
-			Enabled: d.Enabled,
-			Details: d.Details,
-		}
-	}
-
-	return stats
 }
